python-fetcher/requirements.txt
nats-py
aiohttp
asyncio